+++ packages/cx-link/src/components/Link/tokens/cxLink.ts
@@ -13,12 +13,22 @@
  */
 
 import {
-  addProps, Span, addPropsIf, flowIf, as, on, addClassesIf
+  Span,
+  addClassesIf,
+  addProps,
+  addPropsIf,
+  as,
+  flowIf,
+  on,
+  replaceWith,
+  startWith,
+  withProps,
 } from '@bodiless/fclasses';
 import { withSidecarNodes, withNodeKey } from '@bodiless/core';
-import { cxTypography } from '@bodiless/cx-elements';
+import { cxColor, cxTextDecoration, cxTypography } from '@bodiless/cx-elements';
 import { asLinkToken } from '../LinkClean';
 import { useExternalLinkToggle, asEditableLink, useIsDownloadLink } from '../util';
+import { CartIcon } from '../assets/CartIcon';
 
 /**
    * Token which causes link to display as an external link.
@@ -84,6 +94,55 @@ const Sidecar = asLinkToken({
   },
 });
 
+const WhereToBuy = asLinkToken({
+  Components: {
+    Icon: startWith(CartIcon),
+  },
+  Layout: {
+    Wrapper: 'w-full flex justify-center items-center max-w-64 h-12 lg:w-48',
+  },
+  Spacing: {
+    Wrapper: 'mx-auto p-3',
+    Icon: 'mr-3',
+  },
+  Theme: {
+    Wrapper: as(
+      cxColor.BgPrimaryInteractive,
+      cxColor.TextPrimaryFooterCopy,
+      cxTextDecoration.Bold,
+      cxTextDecoration.Uppercase,
+      // @TODO: Create token? It should be same size for both mobile and desktop...
+      'text-m-base',
+      'rounded',
+    ),
+    Icon: 'w-6 h-6',
+    Body: 'leading',
+  },
+  Content: {
+    _: withProps({
+      children: 'Where to Buy',
+    }),
+    Wrapper: withProps({
+      href: '/where-to-buy',
+    }),
+  },
+});
+
+/**
+ * Token that provides the Where To Buy button without an icon.
+ */
+const WhereToBuyWithoutIcon = asLinkToken({
+  ...WhereToBuy,
+  Components: {
+    Icon: replaceWith(() => null),
+  },
+});
+
 export default {
-  Default, WithExternalStyles, WithDownloadStyles, Sidecar,
+  Default,
+  WithExternalStyles,
+  WithDownloadStyles,
+  Sidecar,
+  WhereToBuy,
+  WhereToBuyWithoutIcon,
 };